carried on with the implementation. I need to debug the code and verify its correctness

+++ CppStream/CppStream/include/c_hll.h
@@ -70,7 +70,7 @@ void init_8(hll_8* _hll, unsigned int p, size_t hash_code_len_in_bytes)
 	{
 		_hll->current_sum = 0;
 		_hll->p = p;
-		_hll->m = pow(2, p);
+		_hll->m = (unsigned int)pow(2, p);
 		if (_hll->m == 0)
 		{
 			printf("number overflow for m with a given k value of %d.\n", p);
@@ -88,7 +88,7 @@ void init_8(hll_8* _hll, unsigned int p, size_t hash_code_len_in_bytes)
 		_hll->hash_code_len_in_bits = hash_code_len_in_bytes * CHAR_BIT;
 		_hll->bitmap_offset = _hll->hash_code_len_in_bits - _hll->p - _hll->bitmap_size;
 		_hll->v = _hll->m;
-		_hll->v_size = ceil(_hll->m / 64);
+		_hll->v_size = (size_t)ceil(_hll->m / 64);
 		_hll->v_bitmap = (uint64_t*)calloc(_hll->v_size, sizeof(uint64_t));
 		_hll->condition_value_test_1 = (5 / 2) * _hll->m;
 		_hll->pow_2_to_32 = pow(2, 32);
@@ -114,7 +114,7 @@ void init_32(hll_32* _hll, unsigned int p, size_t hash_code_len_in_bytes)
 	{
 		_hll->current_sum = 0;
 		_hll->p = p;
-		_hll->m = pow(2, p);
+		_hll->m = (unsigned int)pow(2, p);
 		if (_hll->m == 0)
 		{
 			printf("number overflow for m with a given k value of %d.\n", p);
@@ -133,7 +133,7 @@ void init_32(hll_32* _hll, unsigned int p, size_t hash_code_len_in_bytes)
 		_hll->bitmap_offset = (_hll->hash_code_len_in_bits - _hll->p - _hll->bitmap_size) < 0 ? 0 : 
 			(_hll->hash_code_len_in_bits - _hll->p - _hll->bitmap_size);
 		_hll->v = _hll->m;
-		_hll->v_size = ceil(_hll->m / 64);
+		_hll->v_size = (size_t)ceil(_hll->m / 64);
 		_hll->v_bitmap = (uint64_t*)calloc(_hll->v_size, sizeof(uint64_t));
 		_hll->condition_value_test_1 = (5 / 2) * _hll->m;
 		_hll->pow_2_to_32 = pow(2, 32);
@@ -154,7 +154,7 @@ void update_8(hll_8* _hll, uint32_t hash_value)
 	j = hash_value >> (_hll->hash_code_len_in_bits - _hll->p);
 	w = BitWizard::isolate_bits_32(_hll->bitmap_offset, _hll->bitmap_size, hash_value) >> _hll->bitmap_offset;
 	rob = BitWizard::highest_order_bit_index_arch(w);
-	righmost_bit = 1 + log2(rob);
+	righmost_bit = 1 + (uint8_t)log2(rob);
 	current = _hll->bucket[j];
 	_hll->current_sum -= (double)((double)1 / (double)(1 << _hll->bucket[j]));
 	_hll->bucket[j] = BitWizard::max_uint8(current, righmost_bit);
@@ -163,12 +163,12 @@ void update_8(hll_8* _hll, uint32_t hash_value)
 
 void update_8(hll_8* _hll, uint64_t hash_value)
 {
-	uint32_t j;
+	uint64_t j;
 	uint8_t w, current, rob, righmost_bit;
 	j = hash_value >> (_hll->hash_code_len_in_bits - _hll->p);
-	w = BitWizard::isolate_bits_64(_hll->bitmap_offset, _hll->bitmap_size, hash_value) >> _hll->bitmap_offset;
+	w = BitWizard::isolate_bits_64(_hll->bitmap_offset, _hll->bitmap_size, hash_value) >> _hll->bitmap_offset; // this might be a problem
 	rob = BitWizard::highest_order_bit_index_arch(w);
-	righmost_bit = 1 + log2(rob);
+	righmost_bit = 1 + (uint8_t)log2(rob);
 	current = _hll->bucket[j];
 	_hll->current_sum -= (double)((double)1 / (double)(1 << _hll->bucket[j]));
 	_hll->bucket[j] = BitWizard::max_uint8(current, righmost_bit);
@@ -182,18 +182,18 @@ void opt_update_8(hll_8* _hll, uint32_t hash_value)
 	j = hash_value >> (_hll->hash_code_len_in_bits - _hll->p);
 	w = BitWizard::isolate_bits_32(_hll->bitmap_offset, _hll->bitmap_size, hash_value) >> _hll->bitmap_offset;
 	rob = BitWizard::highest_order_bit_index_arch(w);
-	righmost_bit = 1 + log2(rob);
+	righmost_bit = 1 + (uint8_t)log2(rob);
 	current = _hll->bucket[j];
 	_hll->current_sum -= (double)((double)1 / (double)(1 << _hll->bucket[j]));
 	_hll->bucket[j] = BitWizard::max_uint8(current, righmost_bit);
 	_hll->current_sum += (double)((double)1 / (double)(1 << _hll->bucket[j]));
 	if (_hll->bucket[j] != 0)
 	{
-		uint64_t or_operand = 1 << (j % 64);
+		uint64_t or_operand = uint64_t(1 << (j % 64));
 		uint32_t byte_offset = j / 64;
-		uint32_t prev_count = __popcnt64(_hll->v_bitmap[byte_offset]);
+		uint64_t prev_count = __popcnt64(_hll->v_bitmap[byte_offset]);
 		uint64_t new_bitmap_value = _hll->v_bitmap[byte_offset] | or_operand;
-		uint32_t current_count = __popcnt64(new_bitmap_value);
+		uint64_t current_count = __popcnt64(new_bitmap_value);
 		_hll->v_bitmap[byte_offset] = new_bitmap_value;
 		_hll->v -= (current_count - prev_count);
 	}
@@ -201,12 +201,12 @@ void opt_update_8(hll_8* _hll, uint32_t hash_value)
 
 void opt_update_8(hll_8* _hll, uint64_t hash_value)
 {
-	uint32_t j;
+	uint64_t j;
 	uint8_t w, current, rob, righmost_bit;
 	j = hash_value >> (_hll->hash_code_len_in_bits -_hll->p);
-	w = BitWizard::isolate_bits_64(_hll->bitmap_offset, _hll->bitmap_size, hash_value) >> _hll->bitmap_offset;
+	w = BitWizard::isolate_bits_64(_hll->bitmap_offset, _hll->bitmap_size, hash_value) >> _hll->bitmap_offset; // this might be a problem
 	rob = BitWizard::highest_order_bit_index_arch(w);
-	righmost_bit = 1 + log2(rob);
+	righmost_bit = 1 + (uint32_t)log2(rob);
 	current = _hll->bucket[j];
 	_hll->current_sum -= (double)((double)1 / (double)(1 << _hll->bucket[j]));
 	_hll->bucket[j] = BitWizard::max_uint8(current, righmost_bit);
